feat: interface 구현 및 implements 진행
