+++ src/main/java/com/sb02/todoapp/controller/TodoPageController.java
@@ -1,11 +1,7 @@
 package com.sb02.todoapp.controller;
 
-import com.sb02.todoapp.service.TodoService;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 
 @Controller
 public class TodoPageController {
