+++ src/main/java/com/sb02/todoapp/controller/TodoRestController.java
@@ -1,7 +1,7 @@
 package com.sb02.todoapp.controller;
 
 import com.sb02.todoapp.entity.Todo;
-import com.sb02.todoapp.service.TodoService;
+import com.sb02.todoapp.service.TodoServiceImpl;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -12,15 +12,15 @@ import java.util.Map;
 @RestController
 @RequestMapping("/api/todos")
 public class TodoRestController {
-    private final TodoService todoService;
+    private final TodoServiceImpl todoServiceImpl;
 
-    public TodoRestController(TodoService todoService) {
-        this.todoService = todoService;
+    public TodoRestController(TodoServiceImpl todoServiceImpl) {
+        this.todoServiceImpl = todoServiceImpl;
     }
 
     @PostMapping
     public ResponseEntity<Map<String,String>> writeList(@RequestBody TodoCreateDTO todoCreateDTO) {
-        todoService.write(todoCreateDTO.name(), todoCreateDTO.description());
+        todoServiceImpl.write(todoCreateDTO.name(), todoCreateDTO.description());
         Map<String, String> response = new HashMap<>();
         response.put(todoCreateDTO.name(), todoCreateDTO.description());
         return ResponseEntity.ok(response);
@@ -28,7 +28,7 @@ public class TodoRestController {
 
     @GetMapping
     public ResponseEntity<Map<String,List<Todo>>> getTodos() {
-        List<Todo> todo = todoService.getTodo();
+        List<Todo> todo = todoServiceImpl.getTodo();
         Map<String, List<Todo>> response = new HashMap<>();
         response.put("todos", todo);
         return ResponseEntity.ok(response);
